+++ internal/parsergen/backend/golangtable/doc.go
@@ -1,3 +1,0 @@
-// Package golangtable contains the backend which writes the parser as Go source code holding the parse table in lookup
-// tables.
-package golangtable