+++ pkg/parsergen/backend/golang/doc.go
@@ -1,2 +1,3 @@
-// Package golang provides a parser generator backend which takes a parser as input and outputs Go source code.
+// Package golang provides a parser generator backend which takes a parser as input and outputs Go source code
+// which keeps the parse table in lookup tables.
 package golang