+++ pkg/scannergen/backend/golangtable/doc.go
@@ -1,3 +1,0 @@
-// Package golangtable provides a scanner generator backend which takes a DFA as input and outputs Go source code
-// holding the automaton in lookup tables.
-package golangtable